Revert "feat: support template interpolation in multi-agent usage hints (#20973)"
This reverts commit f48b777717e09eb68ef34736d328e96f7a39e9ac.
